requirements.txt
requests
python-dotenv
unittest2
google-generativeai
ipython
absl-py